+++ controllers/organizationController.js
@@ -1,12 +1,21 @@
 const router = require('express').Router();
 const {User, Organization, Incident } = require('../models');
+const { IsUser, IsAuth, IsAdmin } = require("./ValidateUser");
 
 // get all organizations
 // include associated incidents
-router.get('/', (request, response) => {
+router.get('/', async (request, response) => {
+    if(!(await IsUser(request.headers?.authorization?.split(" ")[1]))) {
+        return response.status(403).json({msg:"Invalid access."});
+    }
   Organization.findAll({
     include:[ User ]
   }).then(organizationdata=>{
+    organizationdata.forEach(org => {
+        org.normalCode = "";
+        org.authCode = "";
+        org.adminCode = "";
+    });
     response.json(organizationdata)
 }).catch(error=>{
     console.log(error);
@@ -19,11 +28,19 @@ router.get('/', (request, response) => {
 
 // get one Organization by id value
 // Include associated users
-router.get('/:id', (request, response) => {
+router.get('/:id', async (request, response) => {
+    if(!(await IsUser(request.headers?.authorization?.split(" ")[1]))) {
+        return response.status(403).json({msg:"Invalid access."});
+    }
   Organization.findByPk(request.params.id, {
     include:[User]
-  }).then(organizationdata=>{
+  }).then(async organizationdata=>{
     if(organizationdata){
+        if(!(await IsAdmin(request.headers?.authorization?.split(" ")[1], request.params.id))) {
+            organizationdata.adminCode = "";
+            organizationdata.authCode = "";
+            organizationdata.normalCode = "";
+        }
        return  response.json(organizationdata);
     } else {
       response.status(404).json({
@@ -52,8 +69,11 @@ Organization.create(request.body).then(organizationdata=>{
 })
 });
 
-// update a organization 
-router.put('/:id', (request, response) => {
+// update an organization 
+router.put('/:id', async (request, response) => {
+    if(!(await IsAdmin(request.headers?.authorization?.split(" ")[1]),request.params.id)) {
+        return response.status(403).json({msg:"Invalid access."});
+    }
   // update a Organization's name by its `id` value
     Organization.update({
         name:request.body.name,
@@ -77,7 +97,10 @@ router.put('/:id', (request, response) => {
 });
 
 // delete Organization by its id value
-router.delete('/:id', (request, response) => {
+router.delete('/:id', async (request, response) => {
+    if(!(await IsAdmin(request.headers?.authorization?.split(" ")[1], request.params.id))) {
+        return response.status(403).json({msg:"Invalid access."});
+    }
     Organization.destroy({
     where:{
         id:request.params.id
